refactoring: move generic fixtures to shared location
